refactor: delete unused Navbar and relocate ConnectWallet
